feat: add a Render- and UpdateContext object
Components must pass certain kinds of data to child components during
update and render time. Adding more and more arguments to the update and
draw functions of the component trait is not future proof. The new
RenderContext and UpdateContext struct will now contain all the
necessary for rendering and updating respectively.